\name{MB.pheno}
\alias{MB.pheno}
\docType{data}
\title{
Gene expression for medulloblastoma example
}
\description{
data.frame with gene expression values (rows) for all patients (columns)
}
\usage{}
\source{
Northcott et al. (2011). J Clin Oncol. 29 (11):1408.
}
\references{
Northcott et al. (2011). J Clin Oncol. 29 (11):1408.
}
\examples{
data(MBlastoma)
head(MB.pheno)
}
\keyword{datasets}
