\name{MB.xpr_names}
\alias{MB.xpr_names}
\docType{data}
\title{
Gene names for medulloblastoma expression data

}
\description{
Vector of gene symbols
}
\usage{
}
\source{
Northcott et al. (2011). J Clin Oncol. 29 (11):1408.
}
\references{
Northcott et al. (2011). J Clin Oncol. 29 (11):1408.
}
\examples{
data(MBlastoma)
head(MB.xpr_names)
}
\keyword{datasets}
